+++ app/src/androidTest/java/info/metadude/kotlin/library/roadsigns/demo/DemoActivityParameterizedTest.kt
@@ -3,6 +3,7 @@ package info.metadude.kotlin.library.roadsigns.demo
 import android.content.Context
 import android.view.View
 import androidx.annotation.StringRes
+import androidx.test.espresso.Espresso.onData
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
@@ -10,11 +11,13 @@ import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.rule.ActivityTestRule
 import org.hamcrest.Matcher
+import org.hamcrest.Matchers.*
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
+
 
 @RunWith(Parameterized::class)
 class DemoActivityParameterizedTest(
@@ -41,7 +44,9 @@ class DemoActivityParameterizedTest(
             testCaseOf(R.string.environmental_badges_content_description_red_yellow_green),
             testCaseOf(R.string.diesel_prohibition_cars_free_until_euro_5_v_open_for_residents),
             testCaseOf(R.string.diesel_prohibition_hgvs_free_until_euro_v_open_for_residents_hamburg),
-            testCaseOf(R.string.diesel_prohibition_free_as_of_euro_5_v_except_delivery_vehicles_stuttgart)
+            testCaseOf(R.string.diesel_prohibition_free_as_of_euro_5_v_except_delivery_vehicles_stuttgart),
+            testCaseOf(R.string.prohibition_cars_free_until_euro_5_petrol_until_euro_2_darmstadt),
+            testCaseOf(R.string.prohibition_hgvs_cars_free_until_euro_5_v_petrol_until_euro_2_darmstadt)
         )
 
         private fun testCaseOf(descriptionResource: Int) = arrayOf(
@@ -60,8 +65,12 @@ class DemoActivityParameterizedTest(
     fun contentDescriptionMatchesSelectionText() {
         val selectionItemText = context.getString(selectionItemTextRes)
         val contentDescriptionText = context.getString(contentDescriptionTextRes)
-        onView(withId(R.id.selectionView)).perform(click())
-        onView(withText(selectionItemText)).perform(click())
+        onView(withId(R.id.selectionView))
+            .perform(click())
+        onData(allOf(`is`(instanceOf(String::class.java)), `is`(selectionItemText)))
+            .perform(click())
+        onView(withId(R.id.selectionView))
+            .check(matches(withSpinnerText(containsString(selectionItemText))))
         onView(withId(R.id.genericRoadSignView))
             .check(matches(isDisplayed))
             .check(matches(withContentDescription(contentDescriptionText)))