+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Road signs changelog
 
+## NEXT
+
+* Not released yet.
+
+### Changes
+
+* Add road signs for Diesel prohibition zones in Darmstadt.
+
+
 ## [v.3.0.0](https://github.com/Umweltzone/roadsigns/releases/tag/v.3.0.0)
 
 * Published: 2020-01-21